add settings and docker-compose for postgres
