--- Models/Post.cs
@@ -1,9 +1,0 @@
-namespace EFCoreDemo.Models;
-
-public class Post : EntityBase
-{
-    public string? Title { get; set; }
-    public string? Content { get; set; }
-    public long BlogId { get; set; }
-    public Blog? Blog { get; set; }
-}
