+++ appsettings.json
@@ -6,7 +6,10 @@
     }
   },
   "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "BloggingContext" : "Host=localhost;Database=BloggingDb;Username=SomeUserName;Password=SomePassword"
+  "ConnectionStrings": {},
+  "PostgresDbSettings": {
+    "Host": "localhost",
+    "DbName": "blogs_db",
+    "Username": "postgres"
   }
 }
