+++ docker-compose.yml
@@ -1,0 +1,20 @@
+version: '3.8'
+services:
+  postgres:
+    image: postgres:latest
+    ports:
+      - '5432:5432'
+    environment:
+      - PGDATA=/var/lib/postgres/data/pgdata
+      - POSTGRES_DB=blogs_db
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+    volumes:
+      - pgdata:/var/lib/postgres/data/pgdata
+    networks:
+      - blogs
+networks:
+  blogs:
+    name: blogs
+volumes:
+  pgdata:
